Fix footer link color
Arguably would be nicer to reflect the visited status, but this'll do.

All the module.css we import just gets glommed in as style elements
under the head, so unless we scope it ourselves, it won't actually be
scoped: I can style ALL the a elements on the page from this component.